fps and jpeg quality parser into video settings
